Fix Reset bug
I was forgetting to reset my players array, so there were secretly 4 players
if you played a second round! 😩
This points to a larger problem though, which is that my modules are *stateful*.
I shouldn't *have* to figure out which fields need reseting, I should be able
to just throw the object away and make a new one. Not *quite* worth a full
refactor at this point, but something to bear in mind for the future.